requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
pydantic==2.8.0
pydantic_core==2.20.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
PyYAML==6.0.1
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
wsproto==1.2.0